Edit reducers and actions for deviceList page
