--- src/actions/actionsCreator.js
@@ -1,6 +1,0 @@
-import { CHANGE_STATUS } from '../constants/constants';
-
-export const ChangeStatus = (id) => ({
-  type: CHANGE_STATUS,
-  id
-});
