+++ src/actions/changeStatus/changeStatus.action.js
@@ -1,0 +1,6 @@
+import { CHANGE_STATUS } from '../../constants/constants';
+
+export const changeStatus = (id) => ({
+  type: CHANGE_STATUS,
+  id
+});
