--- src/actions/filterAction.js
@@ -1,8 +1,0 @@
-import { CHANGE_FILTER_OPTION } from '../constants/constants';
-
-export const filterAction = (filterOption) => {
-  return {
-    type: CHANGE_FILTER_OPTION,
-    filterOption
-  };
-};
