--- src/actions/searchAction.js
@@ -1,6 +1,0 @@
-import { SEARCH_ITEM } from '../constants/constants';
-
-export const searchAction = (searchValue) => ({
-  type: SEARCH_ITEM,
-  searchValue
-});
