+++ src/data/componentsNames.js
@@ -18,19 +18,3 @@ export const setingsComponents = {
   Value:ValueSettings,
   Toggle:ToggleSettings
 };
-
-
-export const options = [
-  {
-    value: 'all',
-    innerText: 'All'
-  },
-  {
-    value: 'on',
-    innerText: 'Show On'
-  },
-  {
-    value: 'off',
-    innerText: 'Show OFF'
-  }
-];
