+++ src/data/filterOptions.js
@@ -1,0 +1,14 @@
+export const options = [
+  {
+    value: 'all',
+    innerText: 'All'
+  },
+  {
+    value: 'on',
+    innerText: 'Show On'
+  },
+  {
+    value: 'off',
+    innerText: 'Show OFF'
+  }
+];
