+++ src/reducers/deviceProtoReducer.js
