+++ src/reducers/devicesReducer.js
